+++ test/sql/class.sql
@@ -4,7 +4,7 @@
 SET client_min_messages = warning;
 SET ROLE postgres;
 
-select kind, sql_identifier from pg_ddlx_identify('pg_ddlx_identify(oid)'::regprocedure);
+select kind, sql_identifier from ddlx_identify('ddlx_identify(oid)'::regprocedure);
 
 create function trig() returns trigger as 
 $$begin return old; end $$
@@ -23,7 +23,7 @@ CREATE TABLE test_class_r (
 );
 COMMENT ON TABLE test_class_r IS 'Comment1';
 grant all on test_class_r to public;
-select kind, sql_identifier from pg_ddlx_identify('test_class_r'::regclass);
+select kind, sql_identifier from ddlx_identify('test_class_r'::regclass);
 
 create trigger aaaa before 
 update on test_class_r
@@ -32,10 +32,10 @@ update on test_class_r
 create unique index idx1 on test_class_r (lower(b)) where b is not null;
 create index idx2 on test_class_r using gin (v);
 
-SELECT pg_ddlx_script('test_class_r'::regclass);
-SELECT pg_ddlx_script('test_class_r'::regtype);
-SELECT pg_ddlx_script('idx1'::regclass);
-SELECT pg_ddlx_script('idx2'::regclass);
+SELECT ddlx_script('test_class_r'::regclass);
+SELECT ddlx_script('test_class_r'::regtype);
+SELECT ddlx_script('idx1'::regclass);
+SELECT ddlx_script('idx2'::regclass);
 
 CREATE UNLOGGED TABLE test_class_r2 (
   i  serial, 
@@ -46,22 +46,22 @@ CREATE UNLOGGED TABLE test_class_r2 (
   constraint "blah" foreign key (a) references test_class_r(a) deferrable initially deferred
 );
 alter table test_class_r2 set with oids;
-SELECT pg_ddlx_script('test_class_r2'::regclass);
+SELECT ddlx_script('test_class_r2'::regclass);
 
 CREATE VIEW test_class_v AS
 SELECT * FROM test_class_r 
   WITH CHECK OPTION;
 grant select on test_class_v to public;
-SELECT pg_ddlx_script('test_class_v'::regclass);
-SELECT pg_ddlx_script('test_class_v'::regtype);
+SELECT ddlx_script('test_class_v'::regclass);
+SELECT ddlx_script('test_class_v'::regtype);
 
 CREATE MATERIALIZED VIEW test_class_m AS
 SELECT * FROM test_class_r;
 create unique index test_class_mi ON test_class_m (a);
 
-SELECT pg_ddlx_script('test_class_m'::regclass);
+SELECT ddlx_script('test_class_m'::regclass);
 
-select kind, sql_identifier from pg_ddlx_identify('pg_ddlx_identify(oid)'::regprocedure);
+select kind, sql_identifier from ddlx_identify('ddlx_identify(oid)'::regprocedure);
 
 create function funfun(a int, b text default null, out c numeric, out d text) returns setof record as 
 $$ select 3.14, 'now'::text $$ language sql cost 123 rows 19
@@ -70,19 +70,19 @@ set xmloption = content
 comment on function funfun(int,text) is 'Use more comments!';
 
 select * from funfun(1);
-SELECT pg_ddlx_script('funfun'::regproc);
-SELECT pg_ddlx_script('funfun(int,text)'::regprocedure);
+SELECT ddlx_script('funfun'::regproc);
+SELECT ddlx_script('funfun(int,text)'::regprocedure);
 
 create sequence test_type_S increment 4 start 2;
 comment on sequence test_type_S is 'interleave';
-select pg_ddlx_script('test_type_S'::regclass);
+select ddlx_script('test_type_S'::regclass);
 
 create table test_collation (
 	id serial,
 	c text collate "C" not null,
 	t text
 );
-select pg_ddlx_script('test_collation'::regclass);
+select ddlx_script('test_collation'::regclass);
 
 create view test_class_v_opt1 with (security_barrier) 
 as select * from test_class_v order by 1;
@@ -90,25 +90,25 @@ create view test_class_v_opt2
 as select * from test_class_v order by 1;
 alter  view test_class_v_opt2 set (security_barrier='true');
 
-select pg_ddlx_script('test_class_v_opt1'::regclass);
-select pg_ddlx_script('test_class_v_opt2'::regclass);
+select ddlx_script('test_class_v_opt1'::regclass);
+select ddlx_script('test_class_v_opt2'::regclass);
 
-select pg_ddlx_script('test_class_v_opt2'::regclass::oid);
-select pg_ddlx_script('test_class_v_opt2');
+select ddlx_script('test_class_v_opt2'::regclass::oid);
+select ddlx_script('test_class_v_opt2');
 
 create or replace function test_proc_1() returns text as
 $$ select 'Hello, world!'::text $$ language sql;
 
-select pg_ddlx_script('test_proc_1'::regproc);
-select pg_ddlx_script('test_proc_1'::regproc::oid);
-select pg_ddlx_script('test_proc_1()');
+select ddlx_script('test_proc_1'::regproc);
+select ddlx_script('test_proc_1'::regproc::oid);
+select ddlx_script('test_proc_1()');
 
 CREATE AGGREGATE test_proc_agg_1(text) (
     SFUNC = textcat,
     STYPE = text
 );
 
-select pg_ddlx_script('test_proc_agg_1'::regproc);
+select ddlx_script('test_proc_agg_1'::regproc);
 
 CREATE AGGREGATE test_proc_agg_2(anyelement) (
     SFUNC = array_append,
@@ -116,5 +116,5 @@ CREATE AGGREGATE test_proc_agg_2(anyelement) (
     INITCOND = '{}'
 );
 
-select pg_ddlx_script('test_proc_agg_2'::regproc);
+select ddlx_script('test_proc_agg_2'::regproc);