+++ test/sql/misc.sql
@@ -20,11 +20,11 @@ execute procedure abort_any_command();
 comment on event trigger ddlx_test_event_trigger
      is 'Test event trigger';
 
-select pg_ddlx_create((
+select ddlx_create((
 select oid from pg_event_trigger
  where evtname = 'ddlx_test_event_trigger'));
  
-select pg_ddlx_drop((
+select ddlx_drop((
 select oid from pg_event_trigger
  where evtname = 'ddlx_test_event_trigger'));
  
@@ -34,22 +34,22 @@ CREATE TEXT SEARCH CONFIGURATION english1 ( PARSER = pg_catalog."default" );
 COMMENT ON TEXT SEARCH CONFIGURATION english1 IS 'configuration for english language (1)';
 ALTER TEXT SEARCH CONFIGURATION english1 OWNER TO postgres;
 
-select pg_ddlx_create('english1'::regconfig);
-select pg_ddlx_drop('english1'::regconfig);
+select ddlx_create('english1'::regconfig);
+select ddlx_drop('english1'::regconfig);
 
 CREATE TEXT SEARCH DICTIONARY english1_stem
   ( TEMPLATE = pg_catalog.snowball, language = 'english', stopwords = 'english' );
 COMMENT ON TEXT SEARCH DICTIONARY english1_stem IS 'snowball stemmer for english language (1)';
 ALTER TEXT SEARCH DICTIONARY english1_stem OWNER TO postgres;
 
-select pg_ddlx_create('english1_stem'::regdictionary);
-select pg_ddlx_drop('english1_stem'::regdictionary);
+select ddlx_create('english1_stem'::regdictionary);
+select ddlx_drop('english1_stem'::regdictionary);
 
 CREATE TEXT SEARCH CONFIGURATION simple1 ( PARSER = pg_catalog."default" );
 COMMENT ON TEXT SEARCH CONFIGURATION simple1 IS 'simple configuration (1)';
 ALTER TEXT SEARCH CONFIGURATION simple1 OWNER TO postgres;
 
-select pg_ddlx_create('simple1'::regconfig);
+select ddlx_create('simple1'::regconfig);
 
 CREATE TEXT SEARCH DICTIONARY simple1
   ( TEMPLATE = pg_catalog.simple );
@@ -57,5 +57,5 @@ COMMENT ON TEXT SEARCH DICTIONARY simple1
      IS 'simple dictionary: just lower case and check for stopword (1)';
 ALTER TEXT SEARCH DICTIONARY simple1 OWNER TO postgres;
 
-select pg_ddlx_create('simple1'::regdictionary);
+select ddlx_create('simple1'::regdictionary);