+++ test/sql/script.sql
@@ -1,4 +1,4 @@
 \pset null _null_
 \pset format unaligned
 
-SELECT pg_ddlx_script('test_class_v'::regclass);
+SELECT ddlx_script('test_class_v'::regclass);