+++ test/sql/type.sql
@@ -6,20 +6,20 @@ SET ROLE postgres;
 
 create type test_type_e as enum ('foo','bar','baz','qux');
 comment on type test_type_e is 'my enum';
-select pg_ddlx_script('test_type_e'::regtype);
+select ddlx_script('test_type_e'::regtype);
 
 create domain test_type_d numeric(10,2) check(value is not null) check(value>6) default 7;
 comment on type test_type_d is 'my domain';
-select pg_ddlx_script('test_type_d'::regtype);
+select ddlx_script('test_type_d'::regtype);
 
 create type test_type_c as (i integer, t text, d test_type_d);
 comment on type test_type_c is 'my class type';
 comment on column test_type_c.i is 'my class class column i';
-select pg_ddlx_script('test_type_c'::regtype);
-select pg_ddlx_script('test_type_c'::regclass);
+select ddlx_script('test_type_c'::regtype);
+select ddlx_script('test_type_c'::regclass);
 
 create domain test_collation_d text collate "C" default '***';
-select pg_ddlx_script('test_collation_d'::regtype);
+select ddlx_script('test_collation_d'::regtype);
 
 create type int_t;
 
@@ -64,12 +64,12 @@ CREATE TYPE int_t (
 COMMENT ON TYPE int_t IS '-2 billion to 2 billion integer, 4-byte storage (test)';
 ALTER TYPE int_t OWNER TO postgres;
 
-select pg_ddlx_create('int_t'::regtype);
-select pg_ddlx_create('int_t[]'::regtype);
+select ddlx_create('int_t'::regtype);
+select ddlx_create('int_t[]'::regtype);
 
 /*
-select pg_ddlx_script('daterange'::regtype);
+select ddlx_script('daterange'::regtype);
 
-select pg_ddlx_script('=(integer,integer)'::regoperator);
-select pg_ddlx_script('=(text,text)'::regoperator);
+select ddlx_script('=(integer,integer)'::regoperator);
+select ddlx_script('=(text,text)'::regoperator);
 */